consolidate component folders. resolves #355
