+++ src/components/shell/landingPage/Card.tsx
@@ -9,7 +9,7 @@ import {
   makeStyles,
 } from '@material-ui/core';
 
-import milou from '../../assets/images/milou.png';
+import milou from '../../../assets/images/milou.png';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
